rm DB only when existing

+++ dev/local/restore_db.sh
@@ -22,12 +22,18 @@ elif [ ! -z "$POSTGRES_PASSWORD" ]; then
     if [ ! -z "$POSTGRES_DB" ]; then
         export POSTGRES_NAME=$POSTGRES_DB
     fi
-    echo "Killing clients..."
-    psql -hdb -Upostgres -d$POSTGRES_NAME -c "SELECT pid, (SELECT pg_terminate_backend(pid)) as killed from pg_stat_activity WHERE state LIKE 'idle';"
-    echo "Dropping db..."
-    dropdb -hdb -Upostgres $POSTGRES_NAME
-    echo "Creating new db..."
-    createdb -hdb -Upostgres -T template0 $POSTGRES_NAME
+    if [ "$( psql -hdb -Upostgres -XtAc "SELECT 1 FROM pg_database WHERE datname='$POSTGRES_NAME'" )" = '1' ]
+    then
+        echo "Database already exists"
+        echo "Killing clients..."
+        psql -hdb -Upostgres -d$POSTGRES_NAME -c "SELECT pid, (SELECT pg_terminate_backend(pid)) as killed from pg_stat_activity WHERE state LIKE 'idle';"
+        echo "Dropping db..."
+        dropdb -hdb -Upostgres $POSTGRES_NAME
+    else
+        echo "Database does not exist"
+        echo "Creating new db..."
+        createdb -hdb -Upostgres -T template0 $POSTGRES_NAME
+    fi
     echo "Importing dump..."
     if [[ $FILE == *".gz" ]]; then
         gunzip < $FILE | psql -hdb -Upostgres -d$POSTGRES_NAME
